feat support linux based projects and tests with different conventions

+++ MuTest.Cpp.CLI/Core/CppDirectoryFactory.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using System.Linq;
 using System.Text.RegularExpressions;
 using MuTest.Core.Utility;
 using MuTest.Cpp.CLI.Model;
@@ -46,7 +47,8 @@ namespace MuTest.Cpp.CLI.Core
             var newTestProjectLocation = $"{projectDirectory}\\{newTestProject}";
             var newSolutionLocation = $"{Path.GetDirectoryName(cppClass.TestSolution)}\\{newTestSolution}";
 
-            var solutionCode = Regex.Replace(solution, $"{testProjectName}{testProjectExtension}", newTestProject, RegexOptions.IgnoreCase);
+            var solutionCode = Regex.Replace(solution, $"{testProjectName}{testProjectExtension}", newTestProject,
+                RegexOptions.IgnoreCase);
             solutionCode.UpdateCode(newSolutionLocation);
 
             new FileInfo(cppClass.TestProject).CopyTo(newTestProjectLocation, true);
@@ -79,16 +81,18 @@ namespace MuTest.Cpp.CLI.Core
                     var testContext = new CppTestContext
                     {
                         Index = index
-
                     };
 
                     var newSourceClass = $"{sourceClassName}_mutest_src_{index}{sourceClassExtension}";
                     var newSourceHeader = $"{sourceHeaderName}_mutest_src_{index}{sourceHeaderExtension}";
                     var newTestClass = $"{testClassName}_mutest_test_{index}{testClassExtension}";
 
-                    var testCode = Regex.Replace(test, $"{sourceClassName}{sourceClassExtension}", newSourceClass, RegexOptions.IgnoreCase);
-                    testCode = Regex.Replace(testCode, testClassName, Path.GetFileNameWithoutExtension(newTestClass), RegexOptions.IgnoreCase);
-                    testCode = Regex.Replace(testCode, $"{sourceHeaderName}{sourceHeaderExtension}", newSourceHeader, RegexOptions.IgnoreCase);
+                    var testCode = Regex.Replace(test, $"{sourceClassName}{sourceClassExtension}", newSourceClass,
+                        RegexOptions.IgnoreCase);
+                    testCode = Regex.Replace(testCode, testClassName, Path.GetFileNameWithoutExtension(newTestClass),
+                        RegexOptions.IgnoreCase);
+                    testCode = Regex.Replace(testCode, $"{sourceHeaderName}{sourceHeaderExtension}", newSourceHeader,
+                        RegexOptions.IgnoreCase);
 
                     var newSourceClassLocation = $"{Path.GetDirectoryName(cppClass.SourceClass)}\\{newSourceClass}";
                     var newHeaderClassLocation = $"{Path.GetDirectoryName(cppClass.SourceHeader)}\\{newSourceHeader}";
@@ -101,9 +105,11 @@ namespace MuTest.Cpp.CLI.Core
                     testContext.SourceClass = new FileInfo(newSourceClassLocation);
                     testContext.SourceHeader = new FileInfo(newHeaderClassLocation);
 
-                    if (!sourceHeaderExtension.Equals(sourceClassExtension, StringComparison.InvariantCultureIgnoreCase))
+                    if (!sourceHeaderExtension.Equals(sourceClassExtension,
+                        StringComparison.InvariantCultureIgnoreCase))
                     {
-                        var sourceCode = Regex.Replace(source, $"{sourceHeaderName}{sourceHeaderExtension}", newSourceHeader, RegexOptions.IgnoreCase);
+                        var sourceCode = Regex.Replace(source, $"{sourceHeaderName}{sourceHeaderExtension}",
+                            newSourceHeader, RegexOptions.IgnoreCase);
                         sourceCode.UpdateCode(newSourceClassLocation);
                         new FileInfo(cppClass.SourceHeader).CopyTo(newHeaderClassLocation, true);
 
@@ -217,7 +223,8 @@ namespace MuTest.Cpp.CLI.Core
 
                     var newTestClass = $"{testClassName}_mutest_test_{index}{testClassExtension}";
 
-                    var solutionCode = Regex.Replace(solution, $"{testProjectName}{testProjectExtension}", string.Format(newTestProject, index), RegexOptions.IgnoreCase);
+                    var solutionCode = Regex.Replace(solution, $"{testProjectName}{testProjectExtension}",
+                        string.Format(newTestProject, index), RegexOptions.IgnoreCase);
                     solutionCode.UpdateCode(string.Format(newSolutionLocation, index));
 
                     var newSourceClassLocation = $"{Path.GetDirectoryName(cppClass.SourceClass)}\\{newSourceClass}";
@@ -226,29 +233,37 @@ namespace MuTest.Cpp.CLI.Core
                     var relativeTestCodePath = cppClass.TestClass.RelativePath(projectDirectory);
                     var relativeNewTestCodePath = newTestClassLocation.RelativePath(projectDirectory);
 
-                    var projectXml = Regex.Replace(project, relativeTestCodePath, relativeNewTestCodePath, RegexOptions.IgnoreCase);
-                    projectXml = Regex.Replace(projectXml, $"{sourceClassName}{sourceClassExtension}", newSourceClass, RegexOptions.IgnoreCase);
+                    var projectXml = Regex.Replace(project, relativeTestCodePath, relativeNewTestCodePath,
+                        RegexOptions.IgnoreCase);
+                    projectXml = Regex.Replace(projectXml, $"{sourceClassName}{sourceClassExtension}", newSourceClass,
+                        RegexOptions.IgnoreCase);
                     projectXml.UpdateCode(string.Format(newTestProjectLocation, index));
 
                     newSourceClassLocation.DeleteIfExists();
                     newTestClassLocation.DeleteIfExists();
                     testContext.SourceClass = new FileInfo(newSourceClassLocation);
 
-                    var testCode = Regex.Replace(test, $"{sourceClassName}{sourceClassExtension}", newSourceClass, RegexOptions.IgnoreCase);
+                    var testCode = Regex.Replace(test, $"{sourceClassName}{sourceClassExtension}", newSourceClass,
+                        RegexOptions.IgnoreCase);
 
-                    if (!sourceHeaderExtension.Equals(sourceClassExtension, StringComparison.InvariantCultureIgnoreCase) &&
+                    if (!sourceHeaderExtension.Equals(sourceClassExtension,
+                            StringComparison.InvariantCultureIgnoreCase) &&
                         !string.IsNullOrWhiteSpace(header) &&
                         Regex.IsMatch(header, $"{sourceClassName}{sourceClassExtension}", RegexOptions.IgnoreCase))
                     {
                         var newSourceHeader = $"{sourceHeaderName}_mutest_src_{index}{sourceHeaderExtension}";
-                        var newHeaderClassLocation = $"{Path.GetDirectoryName(cppClass.SourceHeader)}\\{newSourceHeader}";
-                        testCode = Regex.Replace(testCode, $"{sourceHeaderName}{sourceHeaderExtension}", newSourceHeader, RegexOptions.IgnoreCase);
+                        var newHeaderClassLocation =
+                            $"{Path.GetDirectoryName(cppClass.SourceHeader)}\\{newSourceHeader}";
+                        testCode = Regex.Replace(testCode, $"{sourceHeaderName}{sourceHeaderExtension}",
+                            newSourceHeader, RegexOptions.IgnoreCase);
                         newHeaderClassLocation.DeleteIfExists();
 
-                        var sourceCode = Regex.Replace(source, $"{sourceHeaderName}{sourceHeaderExtension}", newSourceHeader, RegexOptions.IgnoreCase);
+                        var sourceCode = Regex.Replace(source, $"{sourceHeaderName}{sourceHeaderExtension}",
+                            newSourceHeader, RegexOptions.IgnoreCase);
                         sourceCode.UpdateCode(newSourceClassLocation);
 
-                        var headerCode = Regex.Replace(header, $"{sourceClassName}{sourceClassExtension}", newSourceClass, RegexOptions.IgnoreCase);
+                        var headerCode = Regex.Replace(header, $"{sourceClassName}{sourceClassExtension}",
+                            newSourceClass, RegexOptions.IgnoreCase);
                         headerCode.UpdateCode(newHeaderClassLocation);
                     }
                     else
@@ -256,13 +271,15 @@ namespace MuTest.Cpp.CLI.Core
                         new FileInfo(cppClass.SourceClass).CopyTo(newSourceClassLocation);
                     }
 
-                    testCode = Regex.Replace(testCode, testClassName, Path.GetFileNameWithoutExtension(newTestClass), RegexOptions.IgnoreCase);
+                    testCode = Regex.Replace(testCode, testClassName, Path.GetFileNameWithoutExtension(newTestClass),
+                        RegexOptions.IgnoreCase);
 
                     testCode.UpdateCode(newTestClassLocation);
                     testContext.TestClass = new FileInfo(newTestClassLocation);
 
                     if (!Regex.IsMatch(testCode, testContext.SourceClass.Name, RegexOptions.IgnoreCase) &&
-                        !Regex.IsMatch(project, $"include.*=.*{sourceClassName}{sourceClassExtension}", RegexOptions.IgnoreCase))
+                        !Regex.IsMatch(project, $"include.*=.*{sourceClassName}{sourceClassExtension}",
+                            RegexOptions.IgnoreCase))
                     {
                         AddSourceReference(testContext);
                     }
@@ -280,6 +297,51 @@ namespace MuTest.Cpp.CLI.Core
             return context;
         }
 
+        public CppBuildContext TakingSourceCodeBackup(CppClass cppClass)
+        {
+            if (cppClass == null)
+            {
+                throw new ArgumentNullException(nameof(cppClass));
+            }
+
+            cppClass.Validate();
+            Reset();
+
+            var sourceClass = new FileInfo(cppClass.SourceClass);
+            var backupSourceClass = new FileInfo($"{sourceClass.FullName}.backup.{DateTime.Now:yyyyMdhhmmss}");
+            var testProject = new FileInfo(cppClass.TestProject);
+            var backupTestProject = new FileInfo($"{testProject.FullName}.backup.{DateTime.Now:yyyyMdhhmmss}");
+            var context = new CppBuildContext
+            {
+                IntDir = "mutest_int_dir/",
+                OutDir = "mutest_out_dir/",
+                IntermediateOutputPath = "mutest_obj_dir/",
+                OutputPath = "mutest_bin_dir/",
+                TestProject = testProject,
+                BackupTestProject = backupTestProject,
+                TestSolution = new FileInfo(cppClass.TestSolution),
+                UseMultipleSolutions = true,
+                TestContexts =
+                {
+                    new CppTestContext
+                    {
+                        Index = 1,
+                        SourceClass = sourceClass,
+                        BackupSourceClass = backupSourceClass,
+                        TestClass = new FileInfo(cppClass.TestClass)
+                    }
+                }
+            };
+
+            sourceClass.CopyTo(backupSourceClass.FullName, true);
+            testProject.CopyTo(backupTestProject.FullName, true);
+
+            DeleteDirectories(context);
+            CreateDirectories(context, 0);
+
+            return context;
+        }
+
         public void DeleteTestFiles(CppBuildContext context)
         {
             if (context == null)
@@ -290,18 +352,36 @@ namespace MuTest.Cpp.CLI.Core
             Reset();
 
             DeleteDirectories(context);
-            context.TestProject.DeleteIfExists();
-            context.TestSolution.DeleteIfExists();
+
+            if (context.TestContexts.FirstOrDefault(x => x.BackupSourceClass != null) == null)
+            {
+                context.TestProject.DeleteIfExists();
+                context.TestSolution.DeleteIfExists();
+            }
+
+            if (context.BackupTestProject != null && File.Exists(context.BackupTestProject.FullName))
+            {
+                context.BackupTestProject.CopyTo(context.TestProject.FullName, true);
+                context.BackupTestProject.DeleteIfExists();
+            }
 
             for (var index = 0; index < context.TestContexts.Count; index++)
             {
                 var testContext = context.TestContexts[index];
-                testContext.SourceClass.DeleteIfExists();
-                testContext.SourceHeader.DeleteIfExists();
-                testContext.TestClass.DeleteIfExists();
+                if (testContext.BackupSourceClass == null)
+                {
+                    testContext.SourceClass.DeleteIfExists();
+                    testContext.SourceHeader.DeleteIfExists();
+                    testContext.TestClass.DeleteIfExists();
 
-                string.Format(context.TestProject.FullName, index).DeleteIfExists();
-                string.Format(context.TestSolution.FullName, index).DeleteIfExists();
+                    string.Format(context.TestProject.FullName, index).DeleteIfExists();
+                    string.Format(context.TestSolution.FullName, index).DeleteIfExists();
+                }
+                else if (File.Exists(testContext.BackupSourceClass.FullName))
+                {
+                    testContext.BackupSourceClass.CopyTo(testContext.SourceClass.FullName, true);
+                    testContext.BackupSourceClass.DeleteIfExists();
+                }
             }
         }
 
@@ -324,10 +404,14 @@ namespace MuTest.Cpp.CLI.Core
             if (testProjectDirectory != null &&
                 testProjectDirectory.Exists)
             {
-                Directory.CreateDirectory(Path.Combine(testProjectDirectory.FullName, string.Format(context.IntDir, index)));
-                Directory.CreateDirectory(Path.Combine(testProjectDirectory.FullName, string.Format(context.OutDir, index)));
-                Directory.CreateDirectory(Path.Combine(testProjectDirectory.FullName, string.Format(context.IntermediateOutputPath, index)));
-                Directory.CreateDirectory(Path.Combine(testProjectDirectory.FullName, string.Format(context.OutputPath, index)));
+                Directory.CreateDirectory(Path.Combine(testProjectDirectory.FullName,
+                    string.Format(context.IntDir, index)));
+                Directory.CreateDirectory(Path.Combine(testProjectDirectory.FullName,
+                    string.Format(context.OutDir, index)));
+                Directory.CreateDirectory(Path.Combine(testProjectDirectory.FullName,
+                    string.Format(context.IntermediateOutputPath, index)));
+                Directory.CreateDirectory(Path.Combine(testProjectDirectory.FullName,
+                    string.Format(context.OutputPath, index)));
             }
         }
 
@@ -359,7 +443,8 @@ namespace MuTest.Cpp.CLI.Core
                         continue;
                     }
 
-                    fileLines.Add($"{Environment.NewLine}#include {start}{testContext.SourceClass.FullName}{end}{Environment.NewLine}");
+                    fileLines.Add(
+                        $"{Environment.NewLine}#include {start}{testContext.SourceClass.FullName}{end}{Environment.NewLine}");
                     fileLines.Add(line);
                     sourceReferenceAdded = true;
                 }
@@ -386,7 +471,8 @@ namespace MuTest.Cpp.CLI.Core
                         continue;
                     }
 
-                    fileLines.Add($"{Environment.NewLine}#include <{testContext.SourceClass.FullName}>{Environment.NewLine}");
+                    fileLines.Add(
+                        $"{Environment.NewLine}#include <{testContext.SourceClass.FullName}>{Environment.NewLine}");
                     fileLines.Add($"namespace mutest_test_{index} {{ {Environment.NewLine}{Environment.NewLine}");
                     fileLines.Add(line);
                     namespaceAdded = true;
